# internal/core/domain/token.go
package domain

type Token struct {
	Access  string `json:"access" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"`
	Refresh string `json:"refresh" example:"57283ca5-6cdd-4751-8839-8193e280fe23"`
}